Move amountField to shared
